+++ examples/physics_avian_controller.rs
@@ -49,11 +49,11 @@ impl TiledPhysicsBackend for MyCustomAvianPhysicsBackend {
         commands: &mut Commands,
         map: &Map,
         collider_source: &TiledColliderSource,
-    ) -> Option<TiledColliderSpawnInfos> {
-        let collider = self.0.spawn_collider(commands, map, collider_source);
-        if let Some(c) = &collider {
+    ) -> Vec<TiledColliderSpawnInfos> {
+        let colliders = self.0.spawn_collider(commands, map, collider_source);
+        for c in &colliders {
             commands.entity(c.entity).insert(RigidBody::Static);
         }
-        collider
+        colliders
     }
 }